Pulls `Coordinate` and `ArtifactRequirements` out into a separate file to avoid a circuilar import later
[ci skip-rust]

[ci skip-build-wheels]
