+++ src/python/pants/jvm/resolve/coursier_fetch.py
@@ -7,16 +7,14 @@ import dataclasses
 import json
 import logging
 import os
-import re
 from dataclasses import dataclass
 from itertools import chain
 from typing import Any, FrozenSet, Iterable, Iterator, List, Tuple
-from urllib.parse import quote_plus as url_quote_plus
 
 from pants.base.glob_match_error_behavior import GlobMatchErrorBehavior
 from pants.core.util_rules.source_files import SourceFiles, SourceFilesRequest
 from pants.engine.addresses import UnparsedAddressInputs
-from pants.engine.collection import Collection, DeduplicatedCollection
+from pants.engine.collection import Collection
 from pants.engine.fs import (
     AddPrefix,
     Digest,
@@ -39,18 +37,16 @@ from pants.jvm.compile import (
     FallibleClasspathEntry,
 )
 from pants.jvm.resolve import coursier_setup
+from pants.jvm.resolve.common import (
+    ArtifactRequirement,
+    ArtifactRequirements,
+    Coordinate,
+    Coordinates,
+)
 from pants.jvm.resolve.coursier_setup import Coursier
 from pants.jvm.resolve.key import CoursierResolveKey
 from pants.jvm.subsystems import JvmSubsystem
-from pants.jvm.target_types import (
-    JvmArtifactArtifactField,
-    JvmArtifactFieldSet,
-    JvmArtifactGroupField,
-    JvmArtifactJarSourceField,
-    JvmArtifactTarget,
-    JvmArtifactUrlField,
-    JvmArtifactVersionField,
-)
+from pants.jvm.target_types import JvmArtifactFieldSet, JvmArtifactJarSourceField, JvmArtifactTarget
 from pants.jvm.util_rules import ExtractFileDigest
 from pants.util.docutil import doc_url
 from pants.util.logging import LogLevel
@@ -82,171 +78,6 @@ class NoCompatibleResolve(Exception):
             f"[compatible_resolves]({doc_url('reference-java_sources#codecompatible_resolvescode')}) "
             "fields) in common."
         )
-
-
-class InvalidCoordinateString(Exception):
-    """The coordinate string being passed is invalid or malformed."""
-
-    def __init__(self, coords: str) -> None:
-        super().__init__(f"Received invalid artifact coordinates: {coords}")
-
-
-@dataclass(frozen=True)
-class Coordinate:
-    """A single Maven-style coordinate for a JVM dependency.
-
-    Coursier uses at least two string serializations of coordinates:
-    1. A format that is accepted by the Coursier CLI which uses trailing attributes to specify
-       optional fields like `packaging`/`type`, `classifier`, `url`, etc. See `to_coord_arg_str`.
-    2. A format in the JSON report, which uses token counts to specify optional fields. We
-       additionally use this format in our own lockfile. See `to_coord_str` and `from_coord_str`.
-    """
-
-    REGEX = re.compile("([^: ]+):([^: ]+)(:([^: ]*)(:([^: ]+))?)?:([^: ]+)")
-
-    group: str
-    artifact: str
-    version: str
-    packaging: str = "jar"
-    classifier: str | None = None
-
-    # True to enforce that the exact declared version of a coordinate is fetched, rather than
-    # allowing dependency resolution to adjust the version when conflicts occur.
-    strict: bool = True
-
-    @staticmethod
-    def from_json_dict(data: dict) -> Coordinate:
-        return Coordinate(
-            group=data["group"],
-            artifact=data["artifact"],
-            version=data["version"],
-            packaging=data.get("packaging", "jar"),
-            classifier=data.get("classifier", None),
-        )
-
-    def to_json_dict(self) -> dict:
-        ret = {
-            "group": self.group,
-            "artifact": self.artifact,
-            "version": self.version,
-            "packaging": self.packaging,
-            "classifier": self.classifier,
-        }
-        return ret
-
-    @classmethod
-    def from_coord_str(cls, s: str) -> Coordinate:
-        """Parses from a coordinate string with optional `packaging` and `classifier` coordinates.
-
-        See the classdoc for more information on the format.
-
-        Using Aether's implementation as reference
-        http://www.javased.com/index.php?source_dir=aether-core/aether-api/src/main/java/org/eclipse/aether/artifact/DefaultArtifact.java
-
-        ${organisation}:${artifact}[:${packaging}[:${classifier}]]:${version}
-
-        See also: `to_coord_str`.
-        """
-
-        parts = Coordinate.REGEX.match(s)
-        if parts is not None:
-            packaging_part = parts.group(4)
-            return cls(
-                group=parts.group(1),
-                artifact=parts.group(2),
-                packaging=packaging_part if packaging_part is not None else "jar",
-                classifier=parts.group(6),
-                version=parts.group(7),
-            )
-        else:
-            raise InvalidCoordinateString(s)
-
-    def as_requirement(self) -> ArtifactRequirement:
-        """Creates a `RequirementCoordinate` from a `Coordinate`."""
-        return ArtifactRequirement(coordinate=self)
-
-    def to_coord_str(self, versioned: bool = True) -> str:
-        """Renders the coordinate in Coursier's JSON-report format, which does not use attributes.
-
-        See also: `from_coord_str`.
-        """
-        unversioned = f"{self.group}:{self.artifact}"
-        if self.classifier is not None:
-            unversioned += f":{self.packaging}:{self.classifier}"
-        elif self.packaging != "jar":
-            unversioned += f":{self.packaging}"
-
-        version_suffix = ""
-        if versioned:
-            version_suffix = f":{self.version}"
-        return f"{unversioned}{version_suffix}"
-
-    def to_coord_arg_str(self, extra_attrs: dict[str, str] | None = None) -> str:
-        """Renders the coordinate in Coursier's CLI input format.
-
-        The CLI input format uses trailing key-val attributes to specify `packaging`, `url`, etc.
-
-        See https://github.com/coursier/coursier/blob/b5d5429a909426f4465a9599d25c678189a54549/modules/coursier/shared/src/test/scala/coursier/parse/DependencyParserTests.scala#L7
-        """
-        attrs = dict(extra_attrs or {})
-        if self.packaging != "jar":
-            # NB: Coursier refers to `packaging` as `type` internally.
-            attrs["type"] = self.packaging
-        if self.classifier:
-            attrs["classifier"] = self.classifier
-        attrs_sep_str = "," if attrs else ""
-        attrs_str = ",".join((f"{k}={v}" for k, v in attrs.items()))
-        return f"{self.group}:{self.artifact}:{self.version}{attrs_sep_str}{attrs_str}"
-
-
-class Coordinates(DeduplicatedCollection[Coordinate]):
-    """An ordered list of `Coordinate`s."""
-
-
-@dataclass(frozen=True)
-class ArtifactRequirement:
-    """A single Maven-style coordinate for a JVM dependency, along with information of how to fetch
-    the dependency if it is not to be fetched from a Maven repository."""
-
-    coordinate: Coordinate
-
-    url: str | None = None
-    jar: JvmArtifactJarSourceField | None = None
-
-    @classmethod
-    def from_jvm_artifact_target(cls, target: Target) -> ArtifactRequirement:
-        if not JvmArtifactFieldSet.is_applicable(target):
-            raise AssertionError(
-                "`ArtifactRequirement.from_jvm_artifact_target()` only works on targets with "
-                "`JvmArtifactFieldSet` fields present."
-            )
-        return ArtifactRequirement(
-            coordinate=Coordinate(
-                group=target[JvmArtifactGroupField].value,
-                artifact=target[JvmArtifactArtifactField].value,
-                version=target[JvmArtifactVersionField].value,
-            ),
-            url=target[JvmArtifactUrlField].value,
-            jar=(
-                target[JvmArtifactJarSourceField]
-                if target[JvmArtifactJarSourceField].value
-                else None
-            ),
-        )
-
-    def to_coord_arg_str(self) -> str:
-        return self.coordinate.to_coord_arg_str(
-            {"url": url_quote_plus(self.url)} if self.url else {}
-        )
-
-
-# TODO: Consider whether to carry classpath scope in some fashion via ArtifactRequirements.
-class ArtifactRequirements(DeduplicatedCollection[ArtifactRequirement]):
-    """An ordered list of Coordinates used as requirements."""
-
-    @classmethod
-    def from_coordinates(cls, coordinates: Iterable[Coordinate]) -> ArtifactRequirements:
-        return ArtifactRequirements(coord.as_requirement() for coord in coordinates)
 
 
 @dataclass(frozen=True)
